new files for login and callback, both working
